+++ CHANGELOG.rst
@@ -7,14 +7,30 @@
 Changelog
 =========
 
-0.6.0 (2024-06-**)
+0.7.0 (2024-07-12)
+------------------
+
+**New features**
+
+* Add optional ``adaptive_clipping`` parameter to :class:`metalearners.DRLearner`.
+
+**Other changes**
+
+* Change the index columns order in ``MetaLearnerGridSearch.results_``.
+
+* Raise a custom error if only one class is present in a classification outcome.
+
+* Raise a custom error if there are some treatment variants which have seen classification outcomes which have not appeared for some other treatment variant.
+
+
+0.6.0 (2024-07-08)
 ------------------
 
 **New features**
 
 * Implement :class:`metalearners.grid_search.MetaLearnerGridSearch`.
 
-* Add ``scoring`` parameter to :meth:`metalearners.metalearner.MetaLearner.evaluate` and
+* Add a ``scoring`` parameter to :meth:`metalearners.metalearner.MetaLearner.evaluate` and
   implement the abstract method for the :class:`metalearners.XLearner` and
   :class:`metalearners.DRLearner`.